resizeBoard w/ scale transform zoom
remade the resizeBoard() function to zoom in and out by changing its transform scale property rather than using height/width because the internet told me to do so if I want to make the zoom follow the cursor (not implemented yet). As per AI suggestion, it also now uses box-shadow (DOM initialization only) for the tile borders instead of actual border lines, since otherwise the lines would look too big when zoomed in, and I can't change the border size to less to 1px. Lastly, the bigger the board, the thinner the border lines too.